backend/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
pydantic==2.8.2
celery==5.4.0
redis==5.0.8
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
scikit-learn==1.5.1
xgboost==2.1.1
torch==2.4.0
qiskit==1.2.0
qiskit-aer==0.15.1
qiskit-ibm-runtime==0.29.0
prometheus-flask-exporter==0.23.0
python-json-logger==2.0.7
marshmallow==3.22.0
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==8.3.2
pytest-cov==5.0.0
